fix(event): refund.completed 컨슈머 트랜잭션 경계 분리
@Transactional 안에서 ack.acknowledge() 를 호출하면 AckMode.MANUAL
환경에서 오프셋이 먼저 커밋된 뒤 DB 커밋이 실패해도 메시지가 재시도되지
않아 처리 기록이 영구 유실될 수 있다.

dedup 체크 + 모니터링 로깅 + markProcessed 를 RefundCompletedService 의
@Transactional 메서드로 분리하고, 컨슈머는 트랜잭션이 커밋된 이후에만
ACK 하도록 PaymentFailedConsumer 와 동일한 패턴으로 정리했다.
